deletecourse컨트롤러

+++ coursemos_spring_boot/src/main/java/com/example/demo/controller/DeleteCourseController.java
@@ -16,6 +16,8 @@ import com.example.demo.domain.Course;
 import com.example.demo.domain.Review;
 import com.example.demo.domain.ReviewLike;
 import com.example.demo.domain.ReviewReadableMember;
+import com.example.demo.service.CompanyService;
+import com.example.demo.service.CouponService;
 import com.example.demo.service.CourseService;
 import com.example.demo.service.PlaceService;
 import com.example.demo.service.ReviewLikeService;
@@ -30,6 +32,10 @@ public class DeleteCourseController {
 	@Autowired
 	private ReviewService reviewService;
 	@Autowired
+	private CouponService couponService;
+	@Autowired
+	private CompanyService companyService;
+	@Autowired
 	private PlaceService placeService;
 	@Autowired
 	private ReviewLikeService reviewLikeService;
@@ -39,31 +45,28 @@ public class DeleteCourseController {
 	
 	@RequestMapping(value="/", method=RequestMethod.POST)
 	public String deleteCourse(@ModelAttribute Course course, BindingResult result, SessionStatus status, HttpServletRequest request) {
-		Review review = reviewService.findReviewById(Long.valueOf(course.getCourseId()));
-		List<ReviewLike> rll = reviewLikeService.findReviewLikeByReviewId(Long.valueOf(course.getCourseId()));
-		if(rll != null) {
-			for(ReviewLike rl : rll) {
-				reviewLikeService.delete(rl);
-			}
-		}
-		List<ReviewReadableMember> rrml= reviewReadableMemberService.findReviewReadableMemberByReviewId(Long.valueOf(course.getCourseId()));
-		if(rrml != null) {
-			for(ReviewReadableMember rrm : rrml) {
-				reviewReadableMemberService.delete(rrm);
-			}
-		}
-		reviewService.delete(review);
 		
-		if (course.getPlace1()!=null) {
-			placeService.deletePlace(course.getPlace1());
-		}
-		if (course.getPlace2()!=null) {
-			placeService.deletePlace(course.getPlace2());
-		}
-		if (course.getPlace3()!=null) {
-			placeService.deletePlace(course.getPlace3());
+
+		List<Review> reviewList = reviewService.findReviewByCourseId(course.getCourseId());
+		if(reviewList != null) {
+			for(Review review : reviewList) {
+				List<ReviewLike> rll = reviewLikeService.findReviewLikeByReviewId(review.getReviewId());
+				if(rll != null) {
+					for(ReviewLike rl : rll) {
+						reviewLikeService.delete(rl);
+					}
+				}
+				List<ReviewReadableMember> rrml= reviewReadableMemberService.findReviewReadableMemberByReviewId(review.getReviewId());
+				if(rrml != null) {
+					for(ReviewReadableMember rrm : rrml) {
+						reviewReadableMemberService.delete(rrm);
+					}
+				}
+				reviewService.delete(review);
+			}
 		}
 
+		
 		courseService.deleteCourse(course);
 		System.out.println(course.getCourseId() + " : delete complete!!!");
